move one tap setup to Boot, follow Sift script loading pattern
